pyproject.toml
[project]
name = "main"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "openagents>=0.4.4",
    "openai>=0.27.10",
    "openai-agents>=0.0.19",
    "python-dotenv>=1.1.1",
]

[project.script]
main = "main:app"
